middleware update

+++ rest/routes/AnswerRoutes.php
@@ -9,7 +9,12 @@ use Firebase\JWT\Key;
  * )
  */
 Flight::route("GET /answers", function(){
+  $user = Flight::get('user');
+  if(isset($user)){
     Flight::json(Flight::answer_service()->get_all());
+  } else {
+    Flight::json(["message" => "User token doesn't exist."], 404);
+  };
 });
 
 /**
@@ -24,7 +29,12 @@ Flight::route("GET /answers", function(){
   * )
   */
 Flight::route("GET /answer_by_id", function(){
+  $user = Flight::get('user');
+  if(isset($user)){
     Flight::json(Flight::answer_service()->get_by_id(Flight::request()->query['id']));
+  } else {
+    Flight::json(["message" => "User token doesn't exist."], 404);
+  };
  });
 
  /**
@@ -34,7 +44,12 @@ Flight::route("GET /answer_by_id", function(){
   * )
   */
 Flight::route("GET /answers/@id", function($id){
+  $user = Flight::get('user');
+  if(isset($user)){
     Flight::json(Flight::answer_service()->get_by_id($id));
+  } else {
+    Flight::json(["message" => "User token doesn't exist."], 404);
+  };
 });
 
 /**
@@ -54,8 +69,13 @@ Flight::route("GET /answers/@id", function($id){
  * )
  */
 Flight::route("DELETE /answers/@id", function($id){
+  $user = Flight::get('user');
+  if(isset($user)){
     Flight::answer_service()->delete($id);
     Flight::json(['message'=>"answer deleted successfully"]);
+  } else {
+    Flight::json(["message" => "User token doesn't exist."], 404);
+  };
 });
 
 /**
@@ -82,11 +102,15 @@ Flight::route("DELETE /answers/@id", function($id){
 * )
 */
 Flight::route("POST /answers", function(){
+  $user = Flight::get('user');
+  if(isset($user)){
     $request= Flight::request()->data->getData();
     Flight::json(['message'=>"Answer added successfully",
                   'data'=>Flight::answer_service()->add($request)
                 ]);
-    
+  } else {
+    Flight::json(["message" => "User token doesn't exist."], 404);
+  };
 });
 
 
@@ -116,12 +140,16 @@ Flight::route("POST /answers", function(){
  */
 //update
 Flight::route("PUT /answers/@id", function($id){
+  $user = Flight::get('user');
+  if(isset($user)){
     $answer = Flight::request()->data->getData();
     #$response=$users_dao->update($user,$id);
     Flight::json(['message'=>"answer edit successfully",
                   'data'=>Flight::answer_service()->update($answer,$id)
                 ]);
-    
+  } else {
+    Flight::json(["message" => "User token doesn't exist."], 404);
+  };
 });
