+++ rest/routes/TypeRoutes.php
@@ -10,7 +10,12 @@ use Firebase\JWT\Key;
  * )
  */
 Flight::route("GET /types", function(){
+  $user = Flight::get('user');
+  if(isset($user)){
     Flight::json(Flight::type_service()->get_all());
+  } else {
+    Flight::json(["message" => "User token doesn't exist."], 404);
+  };
 });
 
 /**
@@ -25,7 +30,12 @@ Flight::route("GET /types", function(){
   * )
   */
 Flight::route("GET /type_by_id", function(){
+  $user = Flight::get('user');
+  if(isset($user)){
     Flight::json(Flight::type_service()->get_by_id(Flight::request()->query['id']));
+  } else {
+    Flight::json(["message" => "User token doesn't exist."], 404);
+  };
  });
 
 /**
@@ -35,7 +45,12 @@ Flight::route("GET /type_by_id", function(){
   * )
   */
 Flight::route("GET /types/@id", function($id){
+  $user = Flight::get('user');
+  if(isset($user)){
     Flight::json(Flight::type_service()->get_by_id($id));
+  } else {
+    Flight::json(["message" => "User token doesn't exist."], 404);
+  };
 });
 
 /**
@@ -55,8 +70,13 @@ Flight::route("GET /types/@id", function($id){
  * )
  */
 Flight::route("DELETE /types/@id", function($id){
+  $user = Flight::get('user');
+  if(isset($user)){
     Flight::type_service()->delete($id);
     Flight::json(['message'=>"type deleted successfully"]);
+  } else {
+    Flight::json(["message" => "User token doesn't exist."], 404);
+  };
 });
 
 /**
@@ -84,11 +104,15 @@ Flight::route("DELETE /types/@id", function($id){
 * )
 */
 Flight::route("POST /types", function(){
+  $user = Flight::get('user');
+  if(isset($user)){
     $request= Flight::request()->data->getData();
     Flight::json(['message'=>"type added successfully",
                   'data'=>Flight::type_service()->add($request)
                 ]);
-    
+  } else {
+    Flight::json(["message" => "User token doesn't exist."], 404);
+  };
 });
 
 /**
@@ -118,12 +142,16 @@ Flight::route("POST /types", function(){
  */
 //update
 Flight::route("PUT /types/@id", function($id){
+  $user = Flight::get('user');
+  if(isset($user)){
     $type = Flight::request()->data->getData();
     #$response=$users_dao->update($user,$id);
     Flight::json(['message'=>"type edit successfully",
                   'data'=>Flight::type_service()->update($type,$id)
                 ]);
-    
+  } else {
+    Flight::json(["message" => "User token doesn't exist."], 404);
+  };
 });
